apps/fake-news-cheeker/requirements.txt
huggingface_hub==0.25.2
transformers>=4.40.0
pandas>=2.2.0
gradio>=4.29.0
sentencepiece
accelerate
scipy
numpy
flask
httpx[http2]==0.24.1
selectolax==0.3.12
beautifulsoup4==4.12.2
readability-lxml==0.8.1
python-dotenv==1.0.0
psycopg2-binary
requests
sqlalchemy
playwright
